test: fix tests failing due to invalid aliases
typed-less-modules is a fork of typed-scss-modules. As such, it inherits some of the code and functionality from typed-scss-modules.

The `--aliases` and `--aliasPrefixes` options, for example, are not actually implemented in typed-less-modules, although the README for typed-less-modules describes how to use these options. However, some tests assume that the `--aliases` and `--aliasPrefixes` options are implemented. Therefore, they are failed.

So here I will stop using alias and let the test pass.